backend/requirements.txt
# Core MCP dependencies
mcp>=1.0.0

# Web framework and ASGI server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# HTTP and middleware
starlette>=0.27.0
aiofiles>=23.0.0

# Web scraping
playwright>=1.40.0

# Authentication
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
PyJWT>=2.8.0

# HTTP client for NHS Terminology Server
aiohttp>=3.9.0

# Database - MySQL Support
PyMySQL>=1.1.0
sqlalchemy>=2.0.0

# Artificial Intelligence and LLM
langfuse
assemblyai
litellm==1.50.4

# Caching and Utilities  
cachetools
pillow
numpy
